objected-oriented -> functional paradigm
1. Updated triangle.py/Triangle and triangulation.py/Triangulation so
that operations on triangles and triangulations uniformly create new
instances rather than modify old ones, in keeping with the functional
style of samsurf.

2. Stylistic changes have been made to the code for readability. These
are minor.

3. Flipping a hinge now has the potential to mutliply points. That is,
if a marked point of one triangle falls on the boundary of two triangles
after a hinge flip, the point is marked on both triangles in the
resulting hinge.