Implemented Triangulation::step_flow
The algorithm computes the straight line trajectory of a specified point
in a specified direction, step-by-step, where to step is to pass from
one triangle to another.

For performance, I have resorted to using floating point arithmetic
because Radical currently does not support arithmetic. However, I worry
about precision errors. We will need to allow for some margin of error
if we are to detect collisions.