New feature: Triangulation:mark_flow
Triangulation:mark_flow will plot the trajectory of a given point under
the straight line flow in a given direction. The motivation for this is
as follows. We would like to be able to determine whether two points lie
on a common trajectory, i.e., whether they illuminate each other. There
is code to perform this check, but it would be nice to be able to
visually confirm this. Triangulation:mark_flow enables this.