OctagonCover
In this branch, I implement the cyclic translation covers of an octagon.
In particular, given a degree and a holonomy representation of the
fundamental group of a genus two surface with six punctures (here
modeled as a degenerate 18-gon, or alternatively, as a regular octagon
with unusual triangulation), OctagonCover yields the corresponding
translation cover as a bowman.triangulation.Triangulation instance.