OctagonCover:flow_permutation
Implemented a method to compute the permutation of a fiber associated to
(1) a point in the base space and (2) a periodic direction. With this,
we can finally begin to gather emperical data and work towards a
solution to the question at hand, which essentially reduces now to: Does
the action of the space of cylinder directions on the fiber take any
point to every other point, or are some pairs of points blocked?