Prym eigenforms of Type A+
Encoded the Prym eigenforms of Type A+, described in Proposition 4.2 and
Figure 3 of Lanneau-Nguyen. To produce a Triangulation instance of one
of these eigenforms, call Triangulation.prym_eigenform_type_aplus with
four integer arguments. These integers are what are denoted in
Proposition 4.2 as w, t, h, and e, in that order.