Bug fix
Prior to this hotfix, the code would enter an infinite loop when
computing permutations of loops that entered the same triangle twice.
This was because it would fail to step the flow in the event that the
trajectory returns to the starting triangle and does not immediately
return to the starting point. As of this commit, this bug is fixed.