Implemented Triangle.intersection
This method returns the intersection set of its collection of marked
lines. A demonstration of this has been appended to the end of the
Jupyter notebook testing_intersection.ipynb. Keep in mind that this
method has yet to be tested thoroughly, but since the bulk of the work
is done by intersect_lines, I think it is safe to use.